handle reconnections to the server when transmitting system resources
