+++ src/packet_transmitter/mod.rs
@@ -1,4 +1,4 @@
-mod dump_dir;
-mod grpc_handler;
+pub(crate) mod dump_dir;
+pub(crate) mod grpc_handler;
 pub(crate) mod packet_buffer;
 pub(crate) mod transmitter;
